+++ composer.json
@@ -15,33 +15,32 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "illuminate/auth": "5.4.*",
-        "illuminate/broadcasting": "5.4.*",
-        "illuminate/bus": "5.4.*",
-        "illuminate/cache": "5.4.*",
-        "illuminate/config": "5.4.*",
-        "illuminate/container": "5.4.*",
-        "illuminate/contracts": "5.4.*",
-        "illuminate/database": "5.4.*",
-        "illuminate/encryption": "5.4.*",
-        "illuminate/events": "5.4.*",
-        "illuminate/filesystem": "5.4.*",
-        "illuminate/hashing": "5.4.*",
-        "illuminate/http": "5.4.*",
-        "illuminate/pagination": "5.4.*",
-        "illuminate/pipeline": "5.4.*",
-        "illuminate/queue": "5.4.*",
-        "illuminate/support": "5.4.*",
-        "illuminate/translation": "5.4.*",
-        "illuminate/validation": "5.4.*",
-        "illuminate/view": "5.4.*",
-        "monolog/monolog": "~1.11",
+        "php": ">=7.0",
+        "illuminate/auth": "5.5.*",
+        "illuminate/broadcasting": "5.5.*",
+        "illuminate/bus": "5.5.*",
+        "illuminate/cache": "5.5.*",
+        "illuminate/config": "5.5.*",
+        "illuminate/container": "5.5.*",
+        "illuminate/contracts": "5.5.*",
+        "illuminate/database": "5.5.*",
+        "illuminate/encryption": "5.5.*",
+        "illuminate/events": "5.5.*",
+        "illuminate/filesystem": "5.5.*",
+        "illuminate/hashing": "5.5.*",
+        "illuminate/http": "5.5.*",
+        "illuminate/pagination": "5.5.*",
+        "illuminate/pipeline": "5.5.*",
+        "illuminate/queue": "5.5.*",
+        "illuminate/support": "5.5.*",
+        "illuminate/translation": "5.5.*",
+        "illuminate/validation": "5.5.*",
+        "illuminate/view": "5.5.*",
+        "monolog/monolog": "~1.12",
         "mtdowling/cron-expression": "~1.0",
-        "nikic/fast-route": "~1.0",
-        "paragonie/random_compat": "~1.4|~2.0",
-        "symfony/http-kernel": "~3.2",
-        "symfony/http-foundation": "~3.2"
+        "nikic/fast-route": "~1.2",
+        "symfony/http-kernel": "~3.3",
+        "symfony/http-foundation": "~3.3"
     },
     "require-dev": {
         "phpunit/phpunit": "~5.7",
@@ -61,7 +60,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.4-dev"
+            "dev-master": "5.5-dev"
         }
     },
     "minimum-stability": "dev",
